requirements.txt
pip>=9.0.1
amqp==1.4.6
gevent==1.2.2
hiredis==0.2.0
PyYAML==3.11
redis==2.10.5
requests[security]==2.7.0
plyvel==0.9
netaddr==0.7.18
antlr4-python2-runtime==4.5.2
jmespath==0.7.1
click==4.1
pan-python==0.10.0
stix==1.1.1.6
cybox==2.1.0.12
libtaxii==1.1.107
pytz==2015.4
certifi
ujson==1.34
filelock==2.0.4
sleekxmpp==1.3.1
beautifulsoup4==4.4.1
cifsdk==2.0.0b7
lz4==0.8.2
networkx==1.11
unicodecsv==0.14.1
blinker==1.4
pyzmq==16.0.3
